Validación de rol y privilegios. Casos de uso retornan texto plano.

+++ src/modules/users/application/use-cases/delete.use-case.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
 import { IUserRepository } from '../../domain/interfaces/user-repository.interface';
 
 @Injectable()
@@ -8,10 +8,17 @@ export class DeleteUserUseCase {
     private readonly userRepository: IUserRepository,
   ) {}
 //se medifico el caso de uso para que no devuelva un void 
-  async execute(id: string): Promise<boolean> {
-    const user = await this.userRepository.findById(id);
-    if (!user) return false;
 
-    return await this.userRepository.delete(id);
+
+  async execute(id: string, userRole: string): Promise<boolean> {
+  if (userRole !== 'admin') {
+    throw new ForbiddenException('Solo un administrador puede eliminar usuarios.');
   }
+
+  const user = await this.userRepository.findById(id);
+  if (!user) return false;
+
+  return await this.userRepository.delete(id);
+}
+
 }
